+++ src/test/java/com/sh/shpay/openbanking/api/OpenBankingControllerTest.java
@@ -1,9 +1,9 @@
 package com.sh.shpay.openbanking.api;
 
 
-import com.sh.shpay.domain.openbanking.openbanking.api.OpenBankingController;
+import com.sh.shpay.domain.openbanking.token.api.OpenBankingTokenController;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 
-@WebMvcTest(OpenBankingController.class)
+@WebMvcTest(OpenBankingTokenController.class)
 public class OpenBankingControllerTest {
 }
